+++ lib/db_core/provider/index.dart
@@ -3,5 +3,6 @@ library;
 export 'dao_providers.dart';
 export 'db_history_provider.dart';
 export 'decrypted_files_guard_provider.dart';
+export 'main_store_backup_orchestrator_provider.dart';
 export 'main_store_provider.dart';
 export 'service_providers.dart';