+++ lib/shared/widgets/titlebar.dart
@@ -4,8 +4,9 @@ import 'package:hoplixi/core/app_prefs/settings_prefs.dart';
 import 'package:hoplixi/core/constants/main_constants.dart';
 import 'package:hoplixi/core/theme/index.dart';
 import 'package:hoplixi/core/utils/toastification.dart';
-import 'package:hoplixi/setup/di_init.dart';
+import 'package:hoplixi/db_core/provider/main_store_backup_orchestrator_provider.dart';
 import 'package:hoplixi/db_core/provider/main_store_provider.dart';
+import 'package:hoplixi/setup/di_init.dart';
 import 'package:hoplixi/shared/widgets/close_database_button.dart';
 import 'package:hoplixi/shared/widgets/language_switcher.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
@@ -60,7 +61,7 @@ class _TitleBarState extends ConsumerState<TitleBar> {
     final scope = _parseBackupScope(scopeRaw);
 
     final result = await ref
-        .read(mainStoreProvider.notifier)
+        .read(mainStoreBackupOrchestratorProvider)
         .createBackup(
           scope: scope,
           outputDirPath: backupPath,